consolidate Popover and Tooltip placement props
